chore: 関数の分離(#8)
